Terminology Section Link and Hover Updates (#201)
Plural versions of terminology terms added where missing, which should
fix the problem of some terms not showing a definition upon hover, as
well as the problem of these same terms not linking back to the
terminology section.

Aggregate Beacons is a great example - the plural versions won't show a
definition on hover, nor will they link to the definition in
terminology, but both of these features work for the singular Aggregate
Beacon, which is what the linking is currently set for in the
Terminology section.

Also added linking for Unsecured BTC1 Update, which was missing. This
should fix the problem with this term not showing on hover or linking to
the terminology section.

**Note - Singleton Beacon and Beacon Type appear twice in the
terminology list, one of each needs to be removed.**